// frontend/types/api/models/StatusEnum.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * * `Inquiry` - Inquiry
 * * `Onboarding` - Onboarding
 * * `Active` - Active
 * * `Churned` - Churned
 */
export enum StatusEnum {
    INQUIRY = 'Inquiry',
    ONBOARDING = 'Onboarding',
    ACTIVE = 'Active',
    CHURNED = 'Churned',
}